move SI-defined constants to si.py
This is a cute logical shuffling-about
(these constants literally DEFINE the SI!)
but also quite useful for CODATA stuff
for a logical import order.
