+++ noether/catalogue/info/spectrum.py
@@ -2,7 +2,7 @@ from ...core.Measure import Measure, MeasureInfo
 from ...core.fundamental import meter
 
 from ..dimensions import length, frequency, energy
-from ..constants import c, h
+from ..si import c, h
 
 SPECTRUM: dict[str, tuple[float, float]] = {
     # ISO 21348. spacewx.com/pdf/SET_21348_2004.pdf
